backend/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.20
requests==2.32.5
pydantic==2.12.4
email-validator==2.3.0
openpyxl==3.1.5
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
